refactor: move fetch logic to app
