+++ src/components/layout.tsx
@@ -1,4 +1,5 @@
 import { ComponentPropsWithoutRef } from "react";
+
 import { cn } from "../lib/utils";
 
 interface LayoutProps extends ComponentPropsWithoutRef<"div"> {
