+++ src/types/error.ts
@@ -1,0 +1,4 @@
+export type CustomError = {
+  title: string;
+  description: string;
+};
